refactor: remove `page` prefix from home related components
