+++ src/components/HomeHero.vue
